Tighten startup channel bootstrap
Verify the current review findings against the branch and address the\nones that still apply.\n\nThis extracts the startup context types into their own module, removes\nthe explicit registrar drop in channel setup, prevents implicit\nwebhook binding to 0.0.0.0, and warns when WASM channels are enabled\nbut the channel directory is missing.
