+++ src/startup/mod.rs
@@ -2,6 +2,9 @@
 
 pub(crate) mod boot;
 pub(crate) mod channels;
+pub(crate) mod context;
 pub(crate) mod phases;
 pub(crate) mod run;
 pub(crate) mod wasm;
+
+pub(crate) use context::*;
